requirements.txt
streamlit
openai
beautifulsoup4
requests
fpdf2
